# machinery/docs/swagger.yaml
basePath: /
definitions:
...
  models.APIResponse:
    properties:
      can_pan_tilt:
        type: boolean
      can_zoom:
        type: boolean
      data: {}
      message: {}
      ptz_functions: {}
    type: object
  models.Authentication:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Authorization:
    properties:
      code:
        type: integer
      expire:
        type: string
      role:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
...
  models.OnvifCredentials:
    properties:
      onvif_password:
        type: string
      onvif_username:
        type: string
      onvif_xaddr:
        type: string
    type: object
  models.OnvifPanTilt:
    properties:
      onvif_credentials:
        $ref: '#/definitions/models.OnvifCredentials'
      pan:
        type: number
      tilt:
        type: number
    type: object
  models.OnvifPreset:
    properties:
      onvif_credentials:
        $ref: '#/definitions/models.OnvifCredentials'
      preset:
        type: string
    type: object
  models.OnvifZoom:
    properties:
      onvif_credentials:
        $ref: '#/definitions/models.OnvifCredentials'
      zoom:
        type: number
    type: object
...
info:
  contact:
    email: support@kerberos.io
    name: API Support
    url: https://www.kerberos.io
  description: This is the API for using and configure Kerberos Agent.
  license:
    name: Apache 2.0 - Commons Clause
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://kerberos.io
  title: Swagger Kerberos Agent API
  version: "1.0"
paths:
  /api/camera/discover:
    get:
      description: Runs an advanced Fing/WiFiman-style scan (ONVIF WS-Discovery +
        TCP port scan + MAC/vendor lookup) and returns the devices found on the local
        network.
      operationId: camera-discover
      parameters:
      - description: Discovery timeout in milliseconds (default 2000)
        in: query
        name: timeout
        type: integer
      - description: Optional subnet(s) to scan, e.g. '192.168.1.0/24' (comma-separated).
          Defaults to the local interfaces.
        in: query
        name: subnet
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Discover cameras and other devices on the local network.
      tags:
      - onvif
  /api/camera/onvif/capabilities:
    post:
      description: Will return the ONVIF capabilities for the specific camera.
      operationId: camera-onvif-capabilities
      parameters:
      - description: OnvifCredentials
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.OnvifCredentials'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Will return the ONVIF capabilities for the specific camera.
      tags:
      - onvif
  /api/camera/onvif/gotopreset:
    post:
      description: Will activate the desired ONVIF preset.
      operationId: camera-onvif-gotopreset
      parameters:
      - description: OnvifPreset
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.OnvifPreset'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Will activate the desired ONVIF preset.
      tags:
      - onvif
  /api/camera/onvif/inputs:
    post:
      description: Will get the digital inputs from the ONVIF device.
      operationId: get-digital-inputs
      parameters:
      - description: OnvifCredentials
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.OnvifCredentials'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - Bearer: []
      summary: Will get the digital inputs from the ONVIF device.
      tags:
      - onvif
  /api/camera/onvif/login:
    post:
      description: Try to login into ONVIF supported camera.
      operationId: camera-onvif-login
      parameters:
      - description: OnvifCredentials
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.OnvifCredentials'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Try to login into ONVIF supported camera.
      tags:
      - onvif
  /api/camera/onvif/outputs:
    post:
      description: Will get the relay outputs from the ONVIF device.
      operationId: get-relay-outputs
      parameters:
      - description: OnvifCredentials
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.OnvifCredentials'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - Bearer: []
      summary: Will get the relay outputs from the ONVIF device.
      tags:
      - onvif
  /api/camera/onvif/outputs/{output}:
    post:
      description: Will trigger the relay output from the ONVIF device.
      operationId: trigger-relay-output
      parameters:
      - description: OnvifCredentials
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.OnvifCredentials'
      - description: Output
        in: path
        name: output
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - Bearer: []
      summary: Will trigger the relay output from the ONVIF device.
      tags:
      - onvif
  /api/camera/onvif/pantilt:
    post:
      description: Panning or/and tilting the camera using a direction (x,y).
      operationId: camera-onvif-pantilt
      parameters:
      - description: OnvifPanTilt
        in: body
        name: panTilt
        required: true
        schema:
          $ref: '#/definitions/models.OnvifPanTilt'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Panning or/and tilting the camera.
      tags:
      - onvif
  /api/camera/onvif/presets:
    post:
      description: Will return the ONVIF presets for the specific camera.
      operationId: camera-onvif-presets
      parameters:
      - description: OnvifCredentials
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.OnvifCredentials'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Will return the ONVIF presets for the specific camera.
      tags:
      - onvif
...
  /api/camera/onvif/zoom:
    post:
      description: Zooming in or out the camera.
      operationId: camera-onvif-zoom
      parameters:
      - description: OnvifZoom
        in: body
        name: zoom
        required: true
        schema:
          $ref: '#/definitions/models.OnvifZoom'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Zooming in or out the camera.
      tags:
      - onvif
...
  /api/login:
    post:
      description: Get Authorization token.
      operationId: login
      parameters:
      - description: Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.Authentication'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Authorization'
      summary: Get Authorization token.
      tags:
      - authentication
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"